Added Edit Functionality to Songs / Touched Up & Fixed
Created the ability to edit a song within the database, created the webpage with it. Added a page for adding a playlist (but has no functionality yet). Also touched up the output for view songs to look clean and take less space. Fixed up errors in ListenerHelper and viewSongs.jsp.